+++ FACTORY/src/main/java/Entities/TypeEnemy.java
@@ -1,5 +1,5 @@
 package Proyecto.FACTORY.src.main.java.Entities;
 public enum TypeEnemy {
     // Is a enum that will help me identify each enemy, and avoid using the SWITCH - CASE
-    ANGELWING_GOLD,ANGELWING_SIVEL,FUNGU_AZUL,FUNGU_RED,FUEGITO,TIERRITA;
+    ANGELWING_GOLD,ANGELWING_SIVEL,ICE,BABY;
 }
