+++ FACTORY/src/main/java/ImplementacionesFungus/Enemy.java
@@ -1,0 +1,8 @@
+package ImplementacionesFungus;
+
+public interface Enemy {
+
+    public void attack();
+
+    public void fend();
+}
