+++ FACTORY/src/main/java/ImplementationsWing/AngeWingGold.java
